Refactor publication code into a separate package
