+++ hpo-case-annotator-core/src/main/java/org/monarchinitiative/hpo_case_annotator/core/publication/PublicationDataParser.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.hpo_case_annotator.core.io;
+package org.monarchinitiative.hpo_case_annotator.core.publication;
 
 import org.monarchinitiative.hpo_case_annotator.model.proto.Publication;
