+++ hpo-case-annotator-core/src/test/java/org/monarchinitiative/hpo_case_annotator/core/publication/PubMedSummaryRetrieverTest.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.hpo_case_annotator.core.io;
+package org.monarchinitiative.hpo_case_annotator.core.publication;
 
 import org.junit.Test;
 import org.monarchinitiative.hpo_case_annotator.core.Utils;
@@ -30,7 +30,7 @@ public class PubMedSummaryRetrieverTest {
      */
     @Test
     public void offlineRetrievalOfPMID() throws Exception {
-        String htmlResponse = Utils.readFile(Paths.get("src/test/resources/org/monarchinitiative/hpo_case_annotator/core/io/pubmed_eutils_summary.xml"));
+        String htmlResponse = Utils.readFile(Paths.get("src/test/resources/org/monarchinitiative/hpo_case_annotator/core/publication/pubmed_eutils_summary.xml"));
 
         Function<String, InputStream> connectionFactory = p -> new ByteArrayInputStream(htmlResponse.getBytes());
         PubMedSummaryRetriever instance = PubMedSummaryRetriever.builder()
