+++ hpo-case-annotator-core/src/test/resources/org/monarchinitiative/hpo_case_annotator/core/publication/pubmed_fmt_epub.txt
